requirements.txt
# Core dependencies for RPG Chatbot
gradio>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced features
pyyaml>=6.0  # For YAML config support (future enhancement)
rich>=13.0.0  # For better console output (optional)

# Development dependencies (optional)
black>=22.0.0  # For code formatting